README: Document x86-64 requirement
Recent versions of the steamwebhelper process
(~/.steam/root/ubuntu12_64/steamwebhelper) are x86-64 executables, and
steamwebhelper is required for a lot of the Steam Client's UI. Many newer
games are also x86-64 executables.

The Steam Client itself and older games continue to be i386 (IA32)
executables, so support for both ABIs is currently necessary.

The 64-bit requirement is difficult to document in terms of Intel brand
names, because most Pentium 4 revisions are i386/IA32-only (more
specifically i686 with SSE2 and sometimes SSE3), but according to
Wikipedia, the Pentium 4 F-series added x86-64/AMD64 support. Pentium D
appears to have been the first Intel brand name that exclusively refers
to x86-64/AMD64 CPUs.

On the AMD side, all Opterons are x86-64/AMD64, so the requirement is
simple to document.

Signed-off-by: Simon McVittie <smcv@collabora.com>